Update to "version 1" and fix things up a bit
I actually may have lost some history here, but now don't want to
try and roll back...

The "new" exposure for DTypes requires to create a static type in C
(not technically, but Python makes it almost impossible to do it
differently).

This is ugly in Cython, but for now I rolled with it, because changing
everything away from cython is also annoying :).